requirements.txt
pillow
Django
django-cleanup
django-cors-headers
djangorestframework
djangorestframework-simplejwt
djoser
mssql-django
gunicorn
whitenoise
dj-database-url
psycopg2
cloudinary 
django-cloudinary-storage

